Doc: remove syntax highlighting guidelines
Since it was decided that selenized will more or less mimic solarized
when it comes to syntax coloring choices, it no longer makes sense to
elaborate on color choices. However, these are valuable notes, so I've
moved them to https://github.com/jan-warchol/limestone-colors (branch
'documentation' as of this moment).